Move roots of unity to utils and ignore vim .swp files
