+++ test-suite/Math/NumberTheory/DirichletCharactersTests.hs
@@ -35,6 +35,7 @@ import Math.NumberTheory.DirichletCharacters
 import qualified Math.NumberTheory.Moduli.Sqrt as J
 import Math.NumberTheory.Moduli.Class (SomeMod(..), modulo)
 import Math.NumberTheory.TestUtils (testSmallAndQuick, Positive(..))
+import Math.NumberTheory.Utils.RootsOfUnity
 
 rootOfUnityTest :: Integer -> Positive Integer -> Bool
 rootOfUnityTest n (Positive d) = toComplex ((d `div` gcd n d) `stimes` toRootOfUnity (n % d)) == (1 :: Complex Double)
